% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_returns}
\alias{create_returns}
\title{Create return series}
\usage{
create_returns(r.mean, r.sd, period)
}
\arguments{
\item{r.mean}{arith mean}

\item{r.sd}{standard deviation}

\item{period}{length of each period}
}
\value{
A series of investment returns
}
\description{
Create return series with time varying mean, sd.
}
