% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_tla}
\alias{get_tla}
\title{Function calculating temporary annuity values from age x to retirment age (fixed end)}
\usage{
get_tla(px, i, scale = rep(1, length(px)))
}
\arguments{
\item{px}{an vector of composite survivial probs from age x to x + n - 1. Length = n}

\item{i}{discount rate, scalar}

\item{scale}{how the annuity scale up over time. eg:
1) salary scale. default is a n vector of 1, meaning no salary scale. used when calculating career based annuity
2) simple COLA scale: COLA increasing at a fixed percentage very year.}
}
\value{
an n vector storing the value of temporary life annuities from age x to age x + n - 1.
}
\description{
suppose the age corresponding to px runs from a1 to aN, and f = aN + 1 (eg. age 30:64, f = 65)
The function computes a..{x, f - x} and s_a..{x, f - x}, x runing from a1 to aN.
The length of px is f - a1
Note that the last element is redundant, just used as a place holder.
}
